Changed to float system, added smart enemy and fixed some bugs... Added FPS counter on screen.

+++ Game/src/com/roget/BasicEnemy.java
@@ -19,7 +19,7 @@ public class BasicEnemy extends GameObject {
 	
 	public Rectangle getBounds()
 	{
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 	public void tick() {
@@ -29,12 +29,12 @@ public class BasicEnemy extends GameObject {
 		if(y <= 0 || y >= Game.HEIGHT - 32) velY *= -1;
 		if(x <= 0 || x >= Game.WIDTH - 16) velX *= -1;
 		
-		handler.addObject(new Trail(x, y, ID.Trail, Color.red, 16, 16, 0.02f, handler));
+		handler.addObject(new Trail((int)x, (int)y, ID.Trail, Color.red, 16, 16, 0.02f, handler));
 	}
 
 	public void render(Graphics g) {
 		g.setColor(Color.red);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 	
 }
