+++ Game/src/com/roget/FastEnemy.java
@@ -19,7 +19,7 @@ public class FastEnemy extends GameObject {
 	
 	public Rectangle getBounds()
 	{
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 	public void tick() {
@@ -34,7 +34,7 @@ public class FastEnemy extends GameObject {
 
 	public void render(Graphics g) {
 		g.setColor(Color.CYAN);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 	
 }
